refactor chart routes
