+++ src/schemas/response.js
@@ -21,4 +21,12 @@ schemas.noContentResponse = createResponseConfig({
     status: { '204': Joi.any().empty() }
 });
 
+schemas.chartResponse = createResponseConfig({
+    schema: Joi.object({
+        id: Joi.string(),
+        title: Joi.string(),
+        metadata: Joi.object()
+    }).unknown()
+});
+
 module.exports = schemas;
